[HBO-6] Create PokemonViewModel class Refactor NewPokemonFragment (add PokemonViewModel) Implement in gradle viewmodel, livedata and fragment
